+++ examples/explorer/data_feed/tycho.rs
@@ -20,19 +20,19 @@ use tycho_core::{dto::Chain, Bytes};
 use tycho_ethereum::BytesCodec;
 use tycho_simulation::{
     evm::{
-        engine_db::{simulation_db::BlockHeader, tycho_db::PreCachedDB},
+        engine_db::{
+            simulation_db::BlockHeader, tycho_db::PreCachedDB, update_engine, SHARED_TYCHO_DB,
+        },
+        protocol::{
+            uniswap_v2::state::UniswapV2State, uniswap_v3::state::UniswapV3State,
+            vm::state::EVMPoolState,
+        },
         tycho_models::{AccountUpdate, ResponseAccount},
     },
     models::ERC20Token,
     protocol::{
         models::{ProtocolComponent, TryFromWithBlock},
         state::ProtocolSim,
-        uniswap_v2::state::UniswapV2State,
-        uniswap_v3::state::UniswapV3State,
-        vm::{
-            engine::{update_engine, SHARED_TYCHO_DB},
-            state::VMPoolState,
-        },
     },
 };
 
@@ -119,7 +119,7 @@ pub async fn process_messages(
     let (jh, mut tycho_stream) = TychoStreamBuilder::new(&tycho_url, Chain::Ethereum)
         .exchange("uniswap_v2", ComponentFilter::with_tvl_range(tvl_threshold, tvl_threshold))
         .exchange("uniswap_v3", ComponentFilter::with_tvl_range(tvl_threshold, tvl_threshold))
-        // .exchange("vm:balancer", ComponentFilter::with_tvl_range(tvl_threshold, tvl_threshold))
+        .exchange("vm:balancer", ComponentFilter::with_tvl_range(tvl_threshold, tvl_threshold))
         .auth_key(auth_key.clone())
         .build()
         .await
@@ -127,7 +127,7 @@ pub async fn process_messages(
 
     let mut all_tokens = load_all_tokens(tycho_url.as_str(), auth_key.as_deref()).await;
 
-    // maps protocols to the the last block we've seen a message for it
+    // maps protocols to the last block we've seen a message for it
     let mut active_protocols: HashMap<String, u64> = HashMap::new();
 
     // persist all protocol states between messages
@@ -225,7 +225,7 @@ pub async fn process_messages(
             {
                 info!("Processing snapshot");
                 let id = Bytes::from_str(&id)
-                    .unwrap_or_else(|_| panic!("Failed parsing H160 from id string {}", id));
+                    .unwrap_or_else(|_| panic!("Failed parsing Bytes from id string {}", id));
                 let mut pair_tokens = Vec::new();
                 let mut skip_pool = false;
 
@@ -285,7 +285,7 @@ pub async fn process_messages(
                             }
                         },
                         "vm:balancer" => {
-                            match VMPoolState::try_from_with_block(
+                            match EVMPoolState::try_from_with_block(
                                 snapshot,
                                 header.clone(),
                                 all_tokens.clone(),
@@ -336,7 +336,7 @@ pub async fn process_messages(
                             let state: &mut Box<dyn ProtocolSim> = entry.get_mut();
                             if let Some(vm_state) = state
                                 .as_any_mut()
-                                .downcast_mut::<VMPoolState<PreCachedDB>>()
+                                .downcast_mut::<EVMPoolState<PreCachedDB>>()
                             {
                                 let tokens: Vec<ERC20Token> = vm_state
                                     .tokens
@@ -360,7 +360,7 @@ pub async fn process_messages(
                                 Some(stored_state) => {
                                     let mut state = stored_state.clone();
                                     if let Some(vm_state) = state.as_any_mut()
-                                        .downcast_mut::<VMPoolState<PreCachedDB>>() {
+                                        .downcast_mut::<EVMPoolState<PreCachedDB>>() {
                                         let tokens: Vec<ERC20Token> = vm_state.tokens
                                             .iter()
                                             .filter_map(|token_address| all_tokens.get(token_address))